chkpt before cleanup
