+++ python_code/eye_analysis/eye_plots/eye_plotter_main.py
@@ -5,18 +5,18 @@ import numpy as np
 from pathlib import Path
 import plotly.graph_objects as go
 
-from python_code.eye_data_cleanup.eye_analysis.plots.plot_dashboard import plot_integrated_dashboard
-from python_code.eye_data_cleanup.eye_analysis.plots.plot_timeseries import plot_pupil_timeseries
-from python_code.eye_data_cleanup.eye_analysis.plots.plot_heatmap import plot_2d_trajectory_heatmap
-from python_code.eye_data_cleanup.eye_analysis.plots.plot_surface_3d import plot_3d_gaze_surface
-from python_code.eye_data_cleanup.eye_analysis.plots.plot_histogram import plot_position_histograms
-from python_code.eye_data_cleanup.eye_viewer import EyeVideoDataset
+from python_code.eye_analysis.eye_plots.plot_dashboard import plot_integrated_dashboard
+from python_code.eye_analysis.eye_plots.plot_timeseries import plot_pupil_timeseries
+from python_code.eye_analysis.eye_plots.plot_heatmap import plot_2d_trajectory_heatmap
+from python_code.eye_analysis.eye_plots.plot_surface_3d import plot_3d_gaze_surface
+from python_code.eye_analysis.eye_plots.plot_histogram import plot_position_histograms
+from python_code.eye_analysis.eye_video_dataset import EyeVideoData
 
 
 class EyeDataPlotter:
     """Analysis and visualization tools for eye tracking data."""
 
-    def __init__(self, *, dataset: EyeVideoDataset) -> None:
+    def __init__(self, *, dataset: EyeVideoData) -> None:
         """Initialize analyzer with eye tracking dataset.
 
         Args:
@@ -30,7 +30,7 @@ class EyeDataPlotter:
         Returns:
             Tuple of (raw_centers, cleaned_centers) as (n_frames, 2) arrays
         """
-        pupil_pairs = [self.dataset.pixel_trajectories.pairs[f'p{i}'] for i in range(1, 9)]
+        pupil_pairs = [self.dataset.dataset.pairs[f'p{i}'] for i in range(1, 9)]
 
         raw_x = np.mean([pair.raw.x for pair in pupil_pairs], axis=0)
         raw_y = np.mean([pair.raw.y for pair in pupil_pairs], axis=0)
@@ -63,14 +63,14 @@ class EyeDataPlotter:
         """
         # Extract frame from video
         video_frame = None
-        if self.dataset.video.video_capture is not None:
-            self.dataset.video.video_capture.set(cv2.CAP_PROP_POS_FRAMES, frame_index)
-            ret, frame = self.dataset.video.video_capture.read()
+        if self.dataset.videos.video_capture is not None:
+            self.dataset.videos.video_capture.set(cv2.CAP_PROP_POS_FRAMES, frame_index)
+            ret, frame = self.dataset.videos.video_capture.read()
             if ret:
                 video_frame = frame
 
         return plot_integrated_dashboard(
-            dataset=self.dataset.pixel_trajectories,
+            dataset=self.dataset.dataset,
             data_name=self.dataset.data_name,
             nbins=nbins,
             show=show,
@@ -94,8 +94,8 @@ class EyeDataPlotter:
             Plotly figure object
         """
         return plot_pupil_timeseries(
-            dataset=self.dataset.pixel_trajectories,
-            frames=self.dataset.pixel_trajectories.frame_indices,
+            dataset=self.dataset.dataset,
+            frames=self.dataset.dataset.frame_indices,
             data_name=self.dataset.data_name,
             show=show,
             output_path=output_path
@@ -121,7 +121,7 @@ class EyeDataPlotter:
             Plotly figure object
         """
         # Get average pupil positions (raw)
-        pupil_pairs = [self.dataset.pixel_trajectories.pairs[f'p{i}'] for i in range(1, 9)]
+        pupil_pairs = [self.dataset.dataset.pairs[f'p{i}'] for i in range(1, 9)]
         x_positions = np.mean([pair.raw.x for pair in pupil_pairs], axis=0)
         y_positions = np.mean([pair.raw.y for pair in pupil_pairs], axis=0)
 
@@ -155,7 +155,7 @@ class EyeDataPlotter:
             Plotly figure object
         """
         # Get average pupil positions (raw)
-        pupil_pairs = [self.dataset.pixel_trajectories.pairs[f'p{i}'] for i in range(1, 9)]
+        pupil_pairs = [self.dataset.dataset.pairs[f'p{i}'] for i in range(1, 9)]
         x_positions = np.mean([pair.raw.x for pair in pupil_pairs], axis=0)
         y_positions = np.mean([pair.raw.y for pair in pupil_pairs], axis=0)
 
@@ -187,7 +187,7 @@ class EyeDataPlotter:
             Plotly figure object
         """
         # Get average pupil positions (raw)
-        pupil_pairs = [self.dataset.pixel_trajectories.pairs[f'p{i}'] for i in range(1, 9)]
+        pupil_pairs = [self.dataset.dataset.pairs[f'p{i}'] for i in range(1, 9)]
         x_positions = np.mean([pair.raw.x for pair in pupil_pairs], axis=0)
         y_positions = np.mean([pair.raw.y for pair in pupil_pairs], axis=0)
 
@@ -259,3 +259,55 @@ class EyeDataPlotter:
         )
 
         print(f"Analysis report saved to: {output_dir}")
+
+
+if __name__ == "__main__":
+    # Setup paths
+    base_path = Path(
+        r"D:\bs\ferret_recordings\2025-07-11_ferret_757_EyeCameras_P43_E15__1\clips\0m_37s-1m_37"
+    )
+    video_path = Path(
+        r"D:\bs\ferret_recordings\2025-07-11_ferret_757_EyeCameras_P43_E15__1\clips\0m_37s-1m_37\eye_data\eye_videos\eye1_clipped_4371_11541.mp4"
+    )
+    timestamps_npy_path = Path(
+        r"D:\bs\ferret_recordings\2025-07-11_ferret_757_EyeCameras_P43_E15__1\clips\0m_37s-1m_37\eye_data\eye_videos\eye1_timestamps_utc_clipped_4371_11541.npy"
+    )
+    csv_path = Path(
+        r"D:\bs\ferret_recordings\2025-07-11_ferret_757_EYeCameras_P43_E15__1\clips\0m_37s-1m_37\eye_data\dlc_output\model_outputs_iteration_11\eye1_clipped_4371_11541DLC_Resnet50_eye_model_v1_shuffle1_snapshot_050.csv"
+    )
+
+    # Create dataset - raw and cleaned data are automatically loaded and filtered
+    print("Loading eye tracking dataset...")
+    eye_dataset = EyeVideoData.create(
+        data_name="ferret_757_eye_tracking",
+        recording_path=base_path,
+        raw_video_path=video_path,
+        timestamps_npy_path=timestamps_npy_path,
+        data_csv_path=csv_path,
+        butterworth_cutoff=6.0,  # Hz - applied to create cleaned data
+        butterworth_sampling_rate=90.0  # Hz (video framerate)
+    )
+
+    # Create analyzer
+    print("Initializing analyzer...")
+    analyzer = EyeDataPlotter(dataset=eye_dataset)
+
+    # Show integrated dashboard with video frame
+    # You can specify which frame to display (default is frame 0)
+    print("\nDisplaying integrated dashboard with video frame...")
+    analyzer.plot_integrated_dashboard(
+        nbins=100,
+        show=True,
+        frame_index=50  # Show frame 50 from the video
+    )
+
+    # # Generate complete analysis report with frame
+    # print("\nGenerating complete analysis report...")
+    # output_dir = base_path / "analysis_output"
+    #
+    # analyzer.create_analysis_report(
+    #     output_dir=output_dir,
+    #     nbins=50,
+    #     frame_index=50  # Include frame 50 in the dashboard
+    # )
+
