+++ python_code/eye_analysis/eye_plots/plot_surface_3d.py
@@ -4,8 +4,8 @@ import numpy as np
 import plotly.graph_objects as go
 from pathlib import Path
 
-from python_code.eye_data_cleanup.eye_analysis.signal_processing import remove_nan_values
-from python_code.eye_data_cleanup.eye_analysis.plots.plot_config import COLORS, get_dark_layout_config, get_dark_scene_config
+from python_code.eye_analysis.data_processing.signal_processing import remove_nan_values
+from python_code.eye_analysis.eye_plots.plot_config import COLORS, get_dark_layout_config, get_dark_scene_config
 
 
 def plot_3d_gaze_surface(
