+++ python_code/eye_analysis/eye_plots/plot_timeseries.py
@@ -5,8 +5,8 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 from pathlib import Path
 
-from python_code.eye_data_cleanup.csv_io import TrajectoryDataset
-from python_code.eye_data_cleanup.eye_analysis.plots.plot_config import (
+from python_code.eye_analysis.trajectory_dataset import TrajectoryDataset
+from python_code.eye_analysis.eye_plots.plot_config import (
     COLORS, get_dark_layout_config, get_dark_axis_config, apply_dark_theme_to_annotations
 )
